Move style modifier tests into seperate file
